core/internal: refactor signal_mutex into its own header/object
