+++ CMakeLists.txt
@@ -759,6 +759,7 @@ add_library (seastar
   src/core/semaphore.cc
   src/core/condition-variable.cc
   src/core/crypto.cc
+  src/core/signal_mutex.cc
   src/http/api_docs.cc
   src/http/common.cc
   src/http/file_handler.cc
